Revises file structure and adds dynamic commands and help documentation.
